Introduce add book dialog and form
- Changed 'add-book' div to 'add-book-button' svg. changed class
to be more descriptive. Changed element type to provide a better
representation of what it does.
- Edited library.updateLibrary() method to accommodate this.
- 'add-book-button' was minimally styled and positioned to provide
basic functionality.
- Added a dialog containing a basic form that appears when the
'add-book-button' is pressed to allow users to add books.
- addBookDialogComponents object was created which contains all
dialog elements and all dialog related functions.
- showDialog(), closeDialog(), and confirmDialog() methods were
added to addBookDialogComponents object which allow users to open
the dialog, close the dialog, and submit the form.
- These functions were attached to eventListeners that can be
initialized using the initializeEventListeners() method on the
addBookDialogComponents object.
- This function as well as library.updateLibrary() are now run at
startup (it is necessary to update the library at startup to ensure
that it contains the relevant books).
- Commented out reference HTML elements now that javascript adds
library elements at startup. I decided to leave the reference HTML
elements commented out to continue referencing for style purposes.