Fix bugs related to form being filled out improperly
- Set maxlength of form inputs to prevent overflow.
- Set word-break on 'book-information' element to prevent overflow.
- Set 'No' to be checked by default when the user is asked if they
have read the book before. This is an alternative way to ensure there
is a value filled out without making the field required.
- Modified book constructor function to check for NaN and empty values
and change the values to 'Unknown'. This was done so that users see
'Unknown' as the book title, author, or number of pages if the field
was filled out improperly or not filled out at all.